# ⚡ Simplify and speed up `thread-data` response parsing